+++ packages/terra-framework-docs/src/terra-dev-site/doc/slide-panel-manager/example/SlidePanelManagerExample.jsx
@@ -194,7 +194,7 @@ class ContentComponent extends React.Component {
           });
         }}
       >
-        {`Disclose (${size})`}
+        {`${size.charAt(0).toUpperCase() + size.slice(1)} panel`}
       </button>
     );
   }
@@ -248,7 +248,7 @@ class ContentComponent extends React.Component {
       <ContentContainer
         header={(
           <ActionHeader
-            title="SlidePanelManager Child Component"
+            text="SlidePanelManager Child Component"
           />
         )}
       >