+++ packages/terra-framework-docs/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Changed
   * Updated `notification-dialog` and `tabs` examples.
+  * Updated examples for `terra-slide-panel-manager`.
 
 ## 1.19.0 - (April 12, 2023)
 
@@ -20,7 +21,7 @@
 * Changed
   * Updated docs for `terra-slide-group`.
   * Updated doc and added test for terra-collapsible-menu-view
-  
+
 * Added
   * Added package-lock.json
   * Added test for `terra-slide-panel`.